GetMissileVel bin exact suggestion #1

+++ Source/missiles.cpp
@@ -680,7 +680,7 @@ void __fastcall GetMissileVel(int i, int sx, int sy, int dx, int dy, int v)
 
 	if (dx != sx || dy != sy) {
 		dxp = (dx + sy - sx - dy) << 21;
-		dyp = (double)((dy + dx - sx - sy) << 21);
+		dyp = (dy + dx - sx - sy) << 21;
 		dr = sqrt(dxp * dxp + dyp * dyp);
 		missile[i]._mixvel = (dxp * (v << 16)) / dr;
 		missile[i]._miyvel = (dyp * (v << 15)) / dr;
